+++ generators/base-application/support/prepare-entity.spec.ts
@@ -20,7 +20,7 @@
 import { beforeEach, describe, it } from 'esmocha';
 import { expect } from 'chai';
 import { formatDateForChangelog } from '../../base/support/index.js';
-import type { EntityAll } from '../../../lib/types/entity-all.js';
+import type { EntityAll } from '../../../lib/types/application-all.d.ts';
 import { entityDefaultConfig, prepareEntityPrimaryKeyForTemplates } from './prepare-entity.js';
 
 describe('generator - base-application - support - prepareEntity', () => {
