+++ generators/base-application/support/task-type-inference.ts
@@ -24,8 +24,7 @@ import type {
   Source as BaseApplicationSource,
 } from '../types.js';
 
-import type { ApplicationAll } from '../../../lib/types/application-properties-all.js';
-import type { EntityAll } from '../../../lib/types/entity-all.js';
+import type { ApplicationAll, EntityAll } from '../../../lib/types/application-all.js';
 import type { SourceAll } from '../../../lib/types/source-all.js';
 
 export function asWriteFilesSection<Data = ApplicationAll<EntityAll>>(section: WriteFileSection<Data>): WriteFileSection<Data> {
